+++ src/WinBridge.Runtime.Contracts/LaunchProcessResultModeValues.cs
@@ -1,0 +1,8 @@
+namespace WinBridge.Runtime.Contracts;
+
+public static class LaunchProcessResultModeValues
+{
+    public const string ProcessStarted = "process_started";
+    public const string ProcessStartedAndExited = "process_started_and_exited";
+    public const string WindowObserved = "window_observed";
+}